moved loc to /components; path refactoring
